requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.9.0
castle==6.1.0
causal-learn==0.1.3.6
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
diskcache==5.6.3
Flask==3.1.0
fonttools==4.54.1
frozenlist==1.5.0
gie==0.0.2
gies==0.0.2
graphviz==0.20.3
guidance==0.2.0rc1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
llguidance==0.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
multidict==6.1.0
networkx==3.2.1
numpy==1.24.4
openai==0.28.0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pyparsing==3.2.0
pyreadstat==1.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
sempler==0.2.13
six==1.16.0
statsmodels==0.14.4
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.66.6
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.17.1
zipp==3.20.2
